Push to Slack based on NewShowingEvent instead
